requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==2.4.0
chardet==4.0.0
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==2.5.0
distlib==0.3.1
dj-database-url==0.5.0
Django
django-cors-headers==3.4.0
django-rest-framework==0.1.0
django-rest-jwt==0.1.46
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djangorestframework-stubs==1.4.0
filelock==3.0.12
html5lib==1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
#Pillow==8.2.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL==19.1.0

pytz==2020.1
#reportlab==3.5.67
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1

txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4

webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
xhtml2pdf==0.2.4
zope.interface==5.1.0
